Save images first
With the other technique, the image description might not be saved by the main xml portion if it was uploaded after we got the title list. This doesn't matter normally if we're saving the image description at the same time as the image, but due to the number of images we're not doing that. Now the only way for an image to not have associated description information is if the image got deleted afterwards, which is unlikely.